requirements.txt
# Auditor de Contratos - RAG + Agente ReAct
# Bootcamp Itaú FIAP 2026

# Core LangChain (versões fixadas para compatibilidade)
langchain>=0.1.0,<0.3.0
langchain-core>=0.1.0,<0.3.0
langchain-openai>=0.0.5,<0.3.0
langchain-community>=0.0.10,<0.3.0

# Vector Database
chromadb>=0.4.22
langchain-chroma>=0.1.0

# Data Processing
pypdf>=3.17.0
python-dotenv>=1.0.0

# Schema Validation
pydantic>=2.5.0

# API
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Optional: Para exercícios avançados
# rank-bm25>=0.2.2  # Para Hybrid Search
# tiktoken>=0.5.0    # Para contar tokens
